requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
dash==2.13.0
dash-bootstrap-components==1.5.0
statsmodels==0.14.0
streamlit==1.28.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0